Pad added
AWESOMEZZZZ

+++ Space-out/Space-out/WinHID.cpp
@@ -39,8 +39,8 @@ void HID::update( UINT message, LPARAM lParam )
 				{
 					POINT t_point = m_mouse.getPosition();
 					Vector2 t_Vpoint;
-					t_Vpoint.x = t_point.x;
-					t_Vpoint.y = t_point.y;
+					t_Vpoint.x = (float)t_point.x;
+					t_Vpoint.y = (float)t_point.y;
 					m_observable.broadcastLeftClick( t_Vpoint );
 				}
 
@@ -48,15 +48,15 @@ void HID::update( UINT message, LPARAM lParam )
 				{
 					POINT t_point = m_mouse.getPosition();
 					Vector2 t_Vpoint;
-					t_Vpoint.x = t_point.x;
-					t_Vpoint.y = t_point.y;
+					t_Vpoint.x = (float)t_point.x;
+					t_Vpoint.y = (float)t_point.y;
 					m_observable.broadcastRightClick(t_Vpoint);
 				}
 
 				POINT t_point = m_mouse.getPosition();
 				Vector2 t_Vpoint;
-				t_Vpoint.x = t_point.x;
-				t_Vpoint.y = t_point.y;
+				t_Vpoint.x = (float)t_point.x;
+				t_Vpoint.y = (float)t_point.y;
 				m_observable.broadcastMousePos(t_Vpoint);
 			}
